Inicia implementação da página de ping para acordar o servidor
